# docs/api/openapi.yaml
openapi: 3.1.0
info:
  title: KBeacon Agent API
  version: 0.1.2
  description: Read-only API for Kubernetes Secret dependency intelligence.
servers:
  - url: http://kbeacon.kbeacon-system.svc:8080
paths:
  /healthz:
    get:
      summary: Liveness endpoint
      operationId: getHealthz
      responses:
        '200':
          description: Agent is alive
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HealthResponse'
  /readyz:
    get:
      summary: Readiness endpoint
      operationId: getReadyz
      responses:
        '200':
          description: Agent is ready
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ReadyResponse'
        '503':
          description: Agent is not ready
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ReadyResponse'
  /api/v1:
    get:
      summary: API discovery
      operationId: getApiDiscovery
      responses:
        '200':
          description: Discovery document
          content:
            application/json:
              schema:
                type: object
                properties:
                  apiVersion:
                    type: string
                  cluster:
                    type: string
                  resources:
                    type: array
                    items:
                      type: string
  /api/v1/secrets:
    get:
      summary: List observed and referenced Secrets
      operationId: listSecrets
      parameters:
        - $ref: '#/components/parameters/Namespace'
        - $ref: '#/components/parameters/OwnerTeam'
        - $ref: '#/components/parameters/Criticality'
        - name: changedSince
          in: query
          schema:
            type: string
          description: Duration such as 15m, 1h, or RFC3339 timestamp.
        - name: minImpactScore
          in: query
          schema:
            type: number
            minimum: 0
            maximum: 100
        - $ref: '#/components/parameters/Limit'
        - $ref: '#/components/parameters/Continue'
      responses:
        '200':
          description: Secret list
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/ListEnvelope'
                  - type: object
                    properties:
                      data:
                        type: array
                        items:
                          $ref: '#/components/schemas/SecretSummary'
        default:
          $ref: '#/components/responses/Error'
  /api/v1/workloads:
    get:
      summary: List normalized workloads
      operationId: listWorkloads
      parameters:
        - $ref: '#/components/parameters/Namespace'
        - $ref: '#/components/parameters/OwnerTeam'
        - $ref: '#/components/parameters/Criticality'
        - name: workloadKind
          in: query
          schema:
            type: string
        - name: includeEdges
          in: query
          schema:
            type: boolean
            default: false
        - $ref: '#/components/parameters/Limit'
        - $ref: '#/components/parameters/Continue'
      responses:
        '200':
          description: Workload list
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/ListEnvelope'
                  - type: object
                    properties:
                      data:
                        type: array
                        items:
                          $ref: '#/components/schemas/WorkloadSummary'
        default:
          $ref: '#/components/responses/Error'
  /api/v1/dependency-map:
    get:
      summary: Get dependency graph slice
      operationId: getDependencyMap
      parameters:
        - $ref: '#/components/parameters/Namespace'
        - $ref: '#/components/parameters/OwnerTeam'
        - name: secretNamespace
          in: query
          schema:
            type: string
        - name: secretName
          in: query
          schema:
            type: string
        - name: includeEdges
          in: query
          schema:
            type: boolean
            default: true
        - $ref: '#/components/parameters/Limit'
      responses:
        '200':
          description: Dependency map
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/Envelope'
                  - type: object
                    properties:
                      data:
                        $ref: '#/components/schemas/DependencyMap'
        default:
          $ref: '#/components/responses/Error'
  /api/v1/secrets/{namespace}/{name}/impact:
    get:
      summary: Get Secret impact
      operationId: getSecretImpact
      parameters:
        - name: namespace
          in: path
          required: true
          schema:
            type: string
        - name: name
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Secret impact details
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/Envelope'
                  - type: object
                    properties:
                      data:
                        $ref: '#/components/schemas/SecretImpact'
        default:
          $ref: '#/components/responses/Error'
  /api/v1/workloads/{namespace}/{name}/dependencies:
    get:
      summary: Get workload dependencies by namespace and name
      operationId: getWorkloadDependenciesByName
      parameters:
        - name: namespace
          in: path
          required: true
          schema:
            type: string
        - name: name
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Workload dependencies
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/Envelope'
                  - type: object
                    properties:
                      data:
                        $ref: '#/components/schemas/WorkloadDependencies'
        default:
          $ref: '#/components/responses/Error'
  /api/v1/workloads/{namespace}/{kind}/{name}/dependencies:
    get:
      summary: Get workload dependencies by namespace, kind, and name
      operationId: getWorkloadDependenciesByKind
      parameters:
        - name: namespace
          in: path
          required: true
          schema:
            type: string
        - name: kind
          in: path
          required: true
          schema:
            type: string
        - name: name
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Workload dependencies
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/Envelope'
                  - type: object
                    properties:
                      data:
                        $ref: '#/components/schemas/WorkloadDependencies'
        default:
          $ref: '#/components/responses/Error'
  /api/v1/teams:
    get:
      summary: List team summaries
      operationId: listTeams
      parameters:
        - $ref: '#/components/parameters/OwnerTeam'
        - $ref: '#/components/parameters/Limit'
        - $ref: '#/components/parameters/Continue'
      responses:
        '200':
          description: Team list
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/ListEnvelope'
                  - type: object
                    properties:
                      data:
                        type: array
                        items:
                          $ref: '#/components/schemas/TeamSummary'
        default:
          $ref: '#/components/responses/Error'
  /api/v1/teams/{team}/impact:
    get:
      summary: Get team impact summary
      operationId: getTeamImpact
      parameters:
        - name: team
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Team impact
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/Envelope'
                  - type: object
                    properties:
                      data:
                        $ref: '#/components/schemas/TeamImpact'
        default:
          $ref: '#/components/responses/Error'
  /api/v1/config:
    get:
      summary: Get sanitized runtime configuration
      operationId: getConfig
      responses:
        '200':
          description: Sanitized configuration
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Envelope'
components:
  parameters:
    Namespace:
      name: namespace
      in: query
      schema:
        type: string
    OwnerTeam:
      name: ownerTeam
      in: query
      schema:
        type: string
    Criticality:
      name: criticality
      in: query
      schema:
        type: string
        enum: [unknown, low, medium, high, critical]
    Limit:
      name: limit
      in: query
      schema:
        type: integer
        default: 100
        minimum: 1
        maximum: 1000
    Continue:
      name: continue
      in: query
      schema:
        type: string
  responses:
    Error:
      description: Error response
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorEnvelope'
  schemas:
    Envelope:
      type: object
      required: [apiVersion, cluster, generatedAt]
      properties:
        apiVersion:
          type: string
          example: kbeacon.io/v1
        cluster:
          type: string
          example: prod-eu-1
        generatedAt:
          type: string
          format: date-time
        data: {}
    ListEnvelope:
      allOf:
        - $ref: '#/components/schemas/Envelope'
        - type: object
          properties:
            pagination:
              $ref: '#/components/schemas/Pagination'
    Pagination:
      type: object
      properties:
        limit:
          type: integer
        continue:
          type: string
        remainingItemCount:
          type: integer
    HealthResponse:
      type: object
      properties:
        status:
          type: string
          example: ok
    ReadyResponse:
      type: object
      properties:
        status:
          type: string
          example: ready
        cluster:
          type: string
        caches:
          type: array
          items:
            type: object
            properties:
              resource:
                type: string
              synced:
                type: boolean
              optional:
                type: boolean
              reason:
                type: string
    SecretRef:
      type: object
      required: [cluster, namespace, name]
      properties:
        cluster:
          type: string
        namespace:
          type: string
        name:
          type: string
        key:
          type: string
    WorkloadRef:
      type: object
      required: [cluster, apiVersion, kind, namespace, name]
      properties:
        cluster:
          type: string
        apiVersion:
          type: string
        kind:
          type: string
        namespace:
          type: string
        name:
          type: string
        uid:
          type: string
    SecretSummary:
      type: object
      properties:
        ref:
          $ref: '#/components/schemas/SecretRef'
        exists:
          type: boolean
        type:
          type: string
        ownerTeam:
          type: string
        criticality:
          type: string
        lastObservedChangeTime:
          type: string
          format: date-time
        observedChangeCount:
          type: integer
          format: uint64
        affectedWorkloadCount:
          type: integer
        affectedTeamCount:
          type: integer
        affectedNamespaceCount:
          type: integer
        impactScore:
          type: number
        unresolvedReferenceCount:
          type: integer
    WorkloadSummary:
      type: object
      properties:
        ref:
          $ref: '#/components/schemas/WorkloadRef'
        ownerTeam:
          type: string
        service:
          type: string
        environment:
          type: string
        criticality:
          type: string
        discoveryMode:
          type: string
          enum: [infer, explicit, hybrid, disabled]
        dependencyCount:
          type: integer
        unresolvedCount:
          type: integer
    DependencySource:
      type: object
      properties:
        type:
          type: string
        path:
          type: string
        container:
          type: string
        initContainer:
          type: string
        ephemeralContainer:
          type: boolean
        volume:
          type: string
        envVar:
          type: string
        annotation:
          type: string
        resourceField:
          type: string
    DependencyEdge:
      type: object
      properties:
        id:
          type: string
        from:
          type: string
        to:
          type: string
        discoveryMode:
          type: string
        sources:
          type: array
          items:
            $ref: '#/components/schemas/DependencySource'
        resolved:
          type: boolean
        optional:
          type: boolean
    DependencyMap:
      type: object
      properties:
        nodes:
          type: array
          items:
            type: object
            properties:
              id:
                type: string
              type:
                type: string
                enum: [workload, secret]
              label:
                type: string
              ref: {}
              ownerTeam:
                type: string
              criticality:
                type: string
              impactScore:
                type: number
        edges:
          type: array
          items:
            $ref: '#/components/schemas/DependencyEdge'
    SecretImpact:
      type: object
      properties:
        secret:
          $ref: '#/components/schemas/SecretSummary'
        summary:
          type: object
          properties:
            affectedWorkloadCount:
              type: integer
            affectedTeamCount:
              type: integer
            affectedNamespaceCount:
              type: integer
            unresolvedReferenceCount:
              type: integer
            discoveryModes:
              type: object
              additionalProperties:
                type: integer
        affectedTeams:
          type: array
          items:
            type: object
            properties:
              ownerTeam:
                type: string
              workloadCount:
                type: integer
        affectedWorkloads:
          type: array
          items:
            $ref: '#/components/schemas/WorkloadSummary'
    WorkloadDependencies:
      type: object
      properties:
        workload:
          $ref: '#/components/schemas/WorkloadSummary'
        dependencies:
          type: array
          items:
            type: object
            properties:
              secret:
                $ref: '#/components/schemas/SecretSummary'
              discoveryMode:
                type: string
              resolved:
                type: boolean
              optional:
                type: boolean
              sources:
                type: array
                items:
                  $ref: '#/components/schemas/DependencySource'
    TeamSummary:
      type: object
      properties:
        ownerTeam:
          type: string
        workloadCount:
          type: integer
        secretCount:
          type: integer
        dependencyCount:
          type: integer
        highImpactSecretCount:
          type: integer
        unresolvedReferenceCount:
          type: integer
    TeamImpact:
      type: object
      properties:
        team:
          $ref: '#/components/schemas/TeamSummary'
        topSecrets:
          type: array
          items:
            $ref: '#/components/schemas/SecretSummary'
        workloads:
          type: array
          items:
            $ref: '#/components/schemas/WorkloadSummary'
    ErrorEnvelope:
      type: object
      properties:
        apiVersion:
          type: string
        error:
          type: object
          properties:
            code:
              type: string
            message:
              type: string
            details:
              type: array
              items:
                type: object